requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
click>=8.1.3
tqdm>=4.66.1
pyyaml>=6.0
langchain>=0.0.312
tiktoken>=0.5.1
